# hackathonApp/backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon App API
  description: |
    REST API for the Hackathon Application. Supports user authentication (with optional
    MFA), ticket management, real-time chat (channels, DMs, ticket threads), Workday
    ad-hoc sync for admin users, and MFA administration.

    ## Authentication
    All authenticated endpoints require a valid session cookie (`hackathon_session`).
    Obtain a session by calling `POST /api/auth/login`. If the user has MFA enabled,
    a two-step login flow is required (see MFA section).

    ## Rate Limiting
    All endpoints are rate-limited. Unauthenticated endpoints are limited per IP;
    authenticated endpoints are limited per user. When rate-limited, the API returns
    `429 Too Many Requests` with a `Retry-After` header.

    ## Error Format
    All errors follow a consistent JSON format:
    ```json
    {
      "error": "error_code_or_message",
      "message": "Human-readable description"
    }
    ```
  version: 1.0.0
  contact:
    name: Hackathon App Team
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://hackathon.example.com/api
    description: Production server
...
paths:
...
  /auth/mfa/enroll:
    post:
      tags: [MFA]
      summary: Start MFA enrollment
      description: Generates a TOTP secret and returns a QR code for the authenticator app.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Enrollment started
          content:
            application/json:
              schema:
                type: object
                properties:
                  qr_code:
                    type: string
                    description: Base64-encoded PNG of QR code
                  manual_key:
                    type: string
                    description: TOTP secret for manual entry
                  issuer:
                    type: string
                  account:
                    type: string
        '400':
          description: MFA already enabled
        '429':
          description: Rate limited
  /auth/mfa/verify-enrollment:
    post:
      tags: [MFA]
      summary: Complete MFA enrollment with TOTP code
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [totp_code]
              properties:
                totp_code:
                  type: string
                  pattern: '^\d{6}$'
      responses:
        '200':
          description: MFA enabled, backup codes returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  backup_codes:
                    type: array
                    items:
                      type: string
                  backup_codes_count:
                    type: integer
                  warning:
                    type: string
        '400':
          description: Invalid TOTP code
  /auth/mfa/verify-login:
    post:
      tags: [MFA]
      summary: Verify TOTP code during login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mfa_token]
              properties:
                mfa_token:
                  type: string
                totp_code:
                  type: string
                  pattern: '^\d{6}$'
                backup_code:
                  type: string
                remember_device:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Login complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccess'
        '400':
          description: Invalid TOTP or backup code
        '401':
          description: Expired mfa_token
        '429':
          description: Rate limited (5 per minute)
  /auth/mfa/backup-codes:
    post:
      tags: [MFA]
      summary: Regenerate backup codes
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        '200':
          description: New backup codes generated
        '401':
          description: Invalid password
  /auth/mfa/disable:
    delete:
      tags: [MFA]
      summary: Disable MFA for current user
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, totp_code]
              properties:
                password:
                  type: string
                totp_code:
                  type: string
      responses:
        '200':
          description: MFA disabled

  # ──────────────────────────────────────────────
  # CHAT
  # ──────────────────────────────────────────────
...
  /admin/mfa/reset/{user_id}:
    post:
      tags: [Admin - MFA]
      summary: Reset MFA for a user (admin only)
      security:
        - sessionCookie: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: MFA reset
        '403':
          description: Admin role required
  /admin/mfa/status:
    get:
      tags: [Admin - MFA]
      summary: Get MFA enrollment statistics
      security:
        - sessionCookie: []
      responses:
        '200':
          description: MFA statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_users:
                    type: integer
                  mfa_enabled:
                    type: integer
                  mfa_disabled:
                    type: integer
                  enforcement:
                    type: string
                  recent_lockouts:
                    type: integer
                  trusted_devices:
                    type: integer
  /admin/mfa/enforcement:
    put:
      tags: [Admin - MFA]
      summary: Set MFA enforcement policy
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [policy]
              properties:
                policy:
                  type: string
                  enum: [optional, required_for_admins, required_for_all]
      responses:
        '200':
          description: Policy updated
        '403':
          description: Admin role required
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hackathon_session
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [hacker, expert, admin]
        workday_id:
          type: string
        department:
          type: string
        table:
          type: string
        team_name:
          type: string
        is_stale:
          type: boolean
        last_synced_at:
          type: string
          format: date-time
          nullable: true
    LoginSuccess:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
...